Further lightmap impl experiments (see desc)
So, none of this is working right. Lightmaps are very weird to me and I fear that the way Source handles them may be incompatible with the static mesh approach I was taking before.

Anyone with more GP experience would be infinitely more equipped to handle this, graphics programming is really not my strong suit